# Conflicts:
#	ClientHandler/src/main/java/ir/sharif/gamein2021/ClientHandler/controller/MainController.java

+++ Core/src/main/java/ir/sharif/gamein2021/core/domain/entity/Team.java
@@ -25,7 +25,7 @@ public class Team implements BaseEntity {
     @Enumerated(value = EnumType.STRING)
     private Country country;
     //TODO this field should be initialize
-    private int credit;
+    private float credit;
 
 
     @Override
